refactor(bin): create a bin entrypoint
