+++ index.js
@@ -2,39 +2,12 @@
 'use strict';
 
 const fs = require('fs');
-const path = require('path');
 const { promisify } = require('util');
 const google = require('googleapis');
-const argv = require('minimist')(process.argv.slice(2));
 const Sherlock = require('sherlockjs');
 const chalk = require('chalk');
 const moment = require('moment');
 const conf = require('./conf');
-const help = require('./help');
-
-/**
- * Get absolute path
- * @param {string} rawPath
- */
-const getPath = (rawPath) => {
-  if (path.isAbsolute(rawPath)) {
-    return rawPath;
-  } else {
-    return path.join(process.cwd(), rawPath);
-  }
-};
-
-/**
- * Error handler
- * @param {Object} err
- */
-const errHandler = (err) => {
-  if (argv.debug) {
-    console.error(`[ERROR] ${err.code} ${err.stack}`);
-  } else {
-    console.error(`[ERROR] ${err.message}`);
-  }
-};
 
 /**
  * Get Oauth2 Client
@@ -61,10 +34,22 @@ const getClient = async () => {
 };
 
 /**
+ * Error handler
+ * @param {Object} err
+ */
+const errHandler = module.exports.errHandler = (err, debug = false) => {
+  if (debug) {
+    console.error(`[ERROR] ${err.code} ${err.stack}`);
+  } else {
+    console.error(`[ERROR] ${err.message}`);
+  }
+};
+
+/**
  * Generate consent page URL
  * @returns {Promise}
  */
-const generateUrl = async () => {
+module.exports.generateUrl = async () => {
   const oauth2Client = getOauth2Client();
   const authUrl = oauth2Client.generateAuthUrl({
     access_type: 'offline',
@@ -78,7 +63,7 @@ const generateUrl = async () => {
  * @param {string} code
  * @returns {Promise}
  */
-const storeToken = async (code) => {
+module.exports.storeToken = async (code) => {
   const oauth2Client = getOauth2Client();
   const tokens = await promisify(oauth2Client.getToken).bind(oauth2Client)(code);
   // oauth2Client.setCredentials(tokens);
@@ -95,7 +80,7 @@ const storeToken = async (code) => {
  * @param {boolean} options.showId
  * @returns {Promise}
  */
-const list = async (naturalInfo, options) => {
+module.exports.list = async (naturalInfo, options) => {
   const { from, to, showId } = options;
   const params = {
     calendarId: conf.CALENDAR_ID,
@@ -156,7 +141,7 @@ const list = async (naturalInfo, options) => {
  * @param {string} options.duration
  * @returns {Promise}
  */
-const insert = async (naturalInfo, options) => {
+module.exports.insert = async (naturalInfo, options) => {
   const event = {};
   if (naturalInfo) {
     const { eventTitle, startDate, endDate, isAllDay } = Sherlock.parse(naturalInfo);
@@ -216,7 +201,7 @@ const insert = async (naturalInfo, options) => {
  * @param {string} eventsPath
  * @returns {Promise}
  */
-const bulk = async (eventsPath) => {
+module.exports.bulk = async (eventsPath) => {
   const events = require(eventsPath);
   const calendar = await getClient();
   const insert = promisify(calendar.events.insert);
@@ -231,64 +216,8 @@ const bulk = async (eventsPath) => {
       });
     } catch (err) {
       console.error(`[ERROR] Error inserting event: ${JSON.stringify(event) || ''}`);
-      errHandler(err);
+      errHandler(err, true);
     }
   });
   await Promise.all(promises);
 };
-
-// main
-(async function () {
-  const command = argv._[0];
-  const configPath = argv.config || argv.C;
-  if (configPath) {
-    const configFile = require(getPath(configPath));
-    Object.assign(conf, configFile);
-  }
-  switch (command) {
-    case 'generateUrl': {
-      await generateUrl();
-      break;
-    }
-    case 'storeToken': {
-      const code = argv._[1];
-      await storeToken(code);
-      break;
-    }
-    case 'list': {
-      const naturalInfo = argv._[1];
-      const params = {
-        from: argv.from || argv.f,
-        to: argv.to || argv.t,
-        showId: argv['show-id'] || argv.i
-      };
-      await list(naturalInfo, params);
-      break;
-    }
-    case 'insert': {
-      const naturalInfo = argv._[1];
-      const params = {
-        summary: argv.summary || argv.s,
-        date: argv.date || argv.d,
-        time: argv.time || argv.t,
-        duration: argv.duration || argv.D
-      };
-      await insert(naturalInfo, params);
-      break;
-    }
-    case 'bulk': {
-      const eventsPath = getPath(argv.events || argv.e);
-      await bulk(eventsPath);
-      break;
-    }
-    case 'help': {
-      console.log(help);
-      break;
-    }
-    default: {
-      console.log('Command not found\n');
-      console.log(help);
-      break;
-    }
-  }
-})().catch(errHandler);
